requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
facebook-sdk==3.1.0
idna==3.10
praw==7.8.1
prawcore==2.4.0
python-dotenv==1.0.1
requests==2.32.3
update-checker==0.18.0
urllib3==2.2.3
websocket-client==1.8.0
